% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saipeAPI.R
\name{saipe}
\alias{saipe}
\alias{saipe_us}
\alias{saipe_state}
\alias{saipe_county}
\title{Get SAIPE data}
\usage{
saipe(geo, year = 2010, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"),
  api_key = NULL)

saipe_us(...)

saipe_state(...)

saipe_county(...)
}
\arguments{
\item{geo}{(character) One of "us", "state", and "county".}

\item{year}{(numeric) A four-digits value or vector of values. Possible values: 1989, 1993, 1995--2014.}

\item{var}{(character) A string or vector of strings. Possible fields: \url{http://api.census.gov/data/timeseries/poverty/saipe/variables.html}}

\item{api_key}{Use \code{saipeAPI::set_api_key("<Your api here>")} to set api_key.}

\item{...}{Pass parameter \code{year} or \code{var} for function \code{saipe_us()}, \code{saipe_state()}, and \code{saipe_county()}. Also see Example section.}
}
\value{
Return a tibble, i.e., modern data frame.
If multiple years are specified, it gives a panel dataset.
}
\description{
Get SAIPE data at various geographic levels.
}
\examples{
\dontrun{
saipeAPI::set_api_key("<Your api here>")

# Get national level median household income and poverty rate in 2014
saipeAPI::saipe(geo = "us", year = 2014, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"))

# Or
saipeAPI::saipe_us(year = 2014, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"))

# Get 2010--2014 data for natinal level, state-level, and county-level
saipeAPI::saipe_us(year = 2010:2014, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"))
saipeAPI::saipe_state(year = 2010:2014, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"))
saipeAPI::saipe_county(year = 2010:2014, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"))
}
}
