% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saipeAPI.R
\name{saipe_us}
\alias{saipe_us}
\title{National level SAIPE data}
\usage{
saipe_us(year = 2010, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"),
  api_key = Sys.getenv("saipe_key"))
}
\arguments{
\item{year}{(numeric) A four-digits value or vector of values. Possible values: 1989, 1993, 1995--2014.}

\item{var}{(character) A string or vector of strings. Possible fields: \url{http://api.census.gov/data/timeseries/poverty/saipe/variables.html}}
}
\value{
Return a tibble. If multiple years are specified, it gives a national level panel dataset.
}
\description{
National level SAIPE data
}
\examples{
\dontrun{
saipeAPI::set_api_key("<Your api here>")

# Get national level median household income and poverty rate in 2014
saipeAPI::saipe_us(year = 2014, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"))

# Get 2010--2014 data
saipeAPI::saipe_us(year = 2010:2014, var = c("NAME", "SAEMHI_PT", "SAEPOVRTALL_PT"))
}
}

