% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saipe_years}
\alias{saipe_years}
\title{Possible years for us, state, and county-level datasets}
\format{An object of class \code{numeric} of length 25.}
\source{
\url{http://www.census.gov/data/developers/data-sets/Poverty-Statistics.html}
}
\usage{
saipe_years
}
\description{
A vector contains all possible years to retrieve
}
\examples{
# To get a vector of possible years:
saipeAPI::saipe_years
}
\keyword{datasets}
