% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saipeAPI.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the Census API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{(character) Get Census's API key at \url{http://api.census.gov/data/key_signup.html}.}
}
\description{
Set the Census API key
}
