% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saipeAPI.R
\name{saipe_state}
\alias{saipe_state}
\title{State level SAIPE data}
\usage{
saipe_state(var = c("SAEMHI_PT", "SAEPOVRTALL_PT"), year = 2010,
  api_key = NULL)
}
\description{
State level SAIPE data
}
\examples{
set_api_key("<Your api here>")
# Get state level median household income and poverty rate in 2014
saipe_state(var = c("SAEMHI_PT", "SAEPOVRTALL_PT"), year = 2014)
}

